Initial project setup with database config
This push is a big one and it essentially just setups up the entire backend. For a in depth view on how to run the project please refer to "docs/env_setup.md".

There are comments on every file to explain what it does and how it works, if there is any confusion let me know.

One thing to note is I coded this on MacOS so I'm not 100% sure if this works on windows. I don't see why it wouldn't but if there are issues again let me know!

TL;DR
Add environment example, gitignore, Docker Compose for PostgreSQL, requirements, and setup documentation. Implement basic database connection logic, main entry point, and test for DB connectivity. Organize project structure with config, database, and parsing modules.