Created a upload FIle script that interact with the databse
This is a script that is able to ask the user if they want to add a file to the database. I created a table in database that stored the file path, the file name, the upload time, and it will store the metadata aswell. (it doesnt yet) I have written tests and they are working aswell)