main.py, and upload_file.py commit.
Currently, we have a few problems that need to be addressed before we can continue pushing any code. I've created a bug-fix branch to do this.

This is a small commit and addresses issues in our main.py, and upload_file.py files.

main.py: The issues addressed were that it was using old SQLite functions instead of the new consentmanager. The imports were inconsistent (redundant) and we were missing file vaildation. Also i got it to initialize data tables.

upload_file.py: This file was missing the table uploaded_files that didn't exist. It essentially was trying to insert nothing. This was simple, I added a new assertion to create that table.